+++ packages/loader/src/GLTFLoader.ts
@@ -1,18 +1,19 @@
 import { AssetPromise, AssetType, Loader, LoadItem, resourceLoader, ResourceManager } from "@galacean/engine-core";
-import { GLTFParser } from "./gltf/GLTFParser";
+import { GLTFPipeline } from "./gltf/GLTFPipeline";
 import { GLTFResource } from "./gltf/GLTFResource";
-import { ParserContext } from "./gltf/parser/ParserContext";
+import { GLTFParserContext } from "./gltf/parser/GLTFParserContext";
 
 @resourceLoader(AssetType.Prefab, ["gltf", "glb"])
 export class GLTFLoader extends Loader<GLTFResource> {
   load(item: LoadItem, resourceManager: ResourceManager): Record<string, AssetPromise<any>> {
-    const url = item.url;
-    const context = new ParserContext(url);
+    const { url } = item;
+    const params = <GLTFParams>item.params;
+    const context = new GLTFParserContext(url);
     const glTFResource = new GLTFResource(resourceManager.engine, url);
     const masterPromiseInfo = context.masterPromiseInfo;
 
     context.glTFResource = glTFResource;
-    context.keepMeshData = item.params?.keepMeshData ?? false;
+    context.keepMeshData = params?.keepMeshData ?? false;
 
     masterPromiseInfo.onCancel(() => {
       const { chainPromises } = context;
@@ -21,8 +22,8 @@ export class GLTFLoader extends Loader<GLTFResource> {
       }
     });
 
-    GLTFParser.defaultPipeline
-      .parse(context)
+    (params?.pipeline || GLTFPipeline.defaultPipeline)
+      ._parse(context)
       .then(masterPromiseInfo.resolve)
       .catch((e) => {
         console.error(e);
@@ -42,4 +43,6 @@ export interface GLTFParams {
    * Keep raw mesh data for glTF parser, default is false.
    */
   keepMeshData: boolean;
+  /** Custom glTF pipeline. */
+  pipeline: GLTFPipeline;
 }
