+++ packages/loader/src/gltf/GLTFSchema.ts
@@ -853,3 +853,6 @@ export interface IGLTF extends IProperty {
    */
   textures?: ITexture[];
 }
+
+/** glTF extensible owner schema. */
+export type GLTFExtensionOwnerSchema = IMeshPrimitive | IMaterial | ITextureInfo | INode;
