+++ packages/loader/src/gltf/extensions/KHR_lights_punctual.ts
@@ -1,13 +1,19 @@
 import { DirectLight, Entity, PointLight, SpotLight } from "@galacean/engine-core";
-import { registerExtension } from "../parser/Parser";
-import { ParserContext } from "../parser/ParserContext";
-import { ExtensionParser } from "./ExtensionParser";
-import { IKHRLightsPunctual_Light } from "./Schema";
+import { registerGLTFExtension } from "../parser/GLTFParser";
+import { GLTFParserContext } from "../parser/GLTFParserContext";
+import { GLTFExtensionMode, GLTFExtensionParser } from "./GLTFExtensionParser";
+import { IKHRLightsPunctual, IKHRLightsPunctual_LightNode } from "./GLTFExtensionSchema";
 
-@registerExtension("KHR_lights_punctual")
-class KHR_lights_punctual extends ExtensionParser {
-  parseEngineResource(schema: IKHRLightsPunctual_Light, entity: Entity, context: ParserContext): void {
-    const { color, intensity = 1, type, range, spot } = schema;
+@registerGLTFExtension("KHR_lights_punctual", GLTFExtensionMode.AdditiveParse)
+class KHR_lights_punctual extends GLTFExtensionParser {
+  /**
+   * @override
+   */
+  additiveParse(context: GLTFParserContext, entity: Entity, extensionSchema: IKHRLightsPunctual_LightNode): void {
+    const lightsSchema = (<IKHRLightsPunctual>context.gltf.extensions.KHR_lights_punctual).lights;
+    const lightSchema = lightsSchema[extensionSchema.light];
+
+    const { color, intensity = 1, type, range, spot } = lightSchema;
     const glTFResource = context.glTFResource;
     let light: DirectLight | PointLight | SpotLight;
