+++ packages/loader/src/gltf/extensions/KHR_materials_clearcoat.ts
@@ -1,13 +1,16 @@
 import { PBRMaterial } from "@galacean/engine-core";
-import { MaterialParser } from "../parser/MaterialParser";
-import { registerExtension } from "../parser/Parser";
-import { ParserContext } from "../parser/ParserContext";
-import { ExtensionParser } from "./ExtensionParser";
-import { IKHRMaterialsClearcoat } from "./Schema";
+import { GLTFMaterialParser } from "../parser/GLTFMaterialParser";
+import { registerGLTFExtension } from "../parser/GLTFParser";
+import { GLTFParserContext } from "../parser/GLTFParserContext";
+import { GLTFExtensionMode, GLTFExtensionParser } from "./GLTFExtensionParser";
+import { IKHRMaterialsClearcoat } from "./GLTFExtensionSchema";
 
-@registerExtension("KHR_materials_clearcoat")
-class KHR_materials_clearcoat extends ExtensionParser {
-  parseEngineResource(schema: IKHRMaterialsClearcoat, material: PBRMaterial, context: ParserContext): void {
+@registerGLTFExtension("KHR_materials_clearcoat", GLTFExtensionMode.AdditiveParse)
+class KHR_materials_clearcoat extends GLTFExtensionParser {
+  /**
+   * @override
+   */
+  additiveParse(context: GLTFParserContext, material: PBRMaterial, schema: IKHRMaterialsClearcoat): void {
     const { textures } = context.glTFResource;
     const {
       clearcoatFactor = 0,
@@ -22,15 +25,15 @@ class KHR_materials_clearcoat extends ExtensionParser {
 
     if (clearcoatTexture) {
       material.clearCoatTexture = textures[clearcoatTexture.index];
-      MaterialParser._parseTextureTransform(material, clearcoatTexture.extensions, context);
+      GLTFMaterialParser._checkOtherTextureTransform(clearcoatTexture, "Clear coat");
     }
     if (clearcoatRoughnessTexture) {
       material.clearCoatRoughnessTexture = textures[clearcoatRoughnessTexture.index];
-      MaterialParser._parseTextureTransform(material, clearcoatRoughnessTexture.extensions, context);
+      GLTFMaterialParser._checkOtherTextureTransform(clearcoatRoughnessTexture, "Clear coat roughness");
     }
     if (clearcoatNormalTexture) {
       material.clearCoatNormalTexture = textures[clearcoatNormalTexture.index];
-      MaterialParser._parseTextureTransform(material, clearcoatNormalTexture.extensions, context);
+      GLTFMaterialParser._checkOtherTextureTransform(clearcoatNormalTexture, "Clear coat normal");
     }
   }
 }
