+++ packages/loader/src/gltf/extensions/KHR_materials_unlit.ts
@@ -1,15 +1,21 @@
 import { UnlitMaterial } from "@galacean/engine-core";
-import { registerExtension } from "../parser/Parser";
-import { ParserContext } from "../parser/ParserContext";
-import { ExtensionParser } from "./ExtensionParser";
-import { IKHRMaterialsUnlit } from "./Schema";
+import { IMaterial } from "../GLTFSchema";
+import { GLTFMaterialParser } from "../parser/GLTFMaterialParser";
+import { registerGLTFExtension } from "../parser/GLTFParser";
+import { GLTFParserContext } from "../parser/GLTFParserContext";
+import { GLTFExtensionMode, GLTFExtensionParser } from "./GLTFExtensionParser";
 
-@registerExtension("KHR_materials_unlit")
-class KHR_materials_unlit extends ExtensionParser {
-  createEngineResource(schema: IKHRMaterialsUnlit, context: ParserContext): UnlitMaterial {
+@registerGLTFExtension("KHR_materials_unlit", GLTFExtensionMode.CreateAndParse)
+class KHR_materials_unlit extends GLTFExtensionParser {
+  /**
+   * @override
+   */
+  createAndParse(context: GLTFParserContext, _, ownerSchema: IMaterial): UnlitMaterial {
     const { engine } = context.glTFResource;
     const material = new UnlitMaterial(engine);
+    material.name = ownerSchema.name;
 
+    GLTFMaterialParser._parseStandardProperty(context, material, ownerSchema);
     return material;
   }
 }
