+++ packages/loader/src/gltf/extensions/index.ts
@@ -6,10 +6,13 @@ import "./KHR_materials_pbrSpecularGlossiness";
 import "./KHR_materials_sheen";
 import "./KHR_materials_transmission";
 import "./KHR_materials_unlit";
-import "./KHR_materials_variants";
+export type { IGLTFExtensionVariants } from "./KHR_materials_variants";
 import "./KHR_materials_volume";
 import "./KHR_mesh_quantization";
 import "./KHR_texture_basisu";
 import "./KHR_texture_transform";
 import "./GALACEAN_materials_remap";
 import "./KHR_materials_ior";
+
+export { GLTFExtensionParser, GLTFExtensionMode } from "./GLTFExtensionParser";
+export * from "./GLTFExtensionSchema";
