+++ packages/loader/src/gltf/parser/GLTFBufferParser.ts
@@ -1,11 +1,11 @@
 import { AssetPromise, request } from "@galacean/engine-core";
 import { GLTFUtil } from "../GLTFUtil";
-import { IBuffer, IGLTF } from "../Schema";
-import { Parser } from "./Parser";
-import { ParserContext } from "./ParserContext";
+import { IBuffer, IGLTF } from "../GLTFSchema";
+import { GLTFParser } from "./GLTFParser";
+import { GLTFParserContext } from "./GLTFParserContext";
 
-export class BufferParser extends Parser {
-  parse(context: ParserContext): AssetPromise<void> {
+export class GLTFBufferParser extends GLTFParser {
+  parse(context: GLTFParserContext): AssetPromise<void> {
     const glTFResource = context.glTFResource;
     const { url } = glTFResource;
