+++ packages/loader/src/index.ts
@@ -2,7 +2,6 @@ import "./AnimatorControllerLoader";
 import "./BufferLoader";
 import "./EnvLoader";
 import "./FontLoader";
-import "./gltf/extensions/index";
 import "./GLTFLoader";
 import "./HDRLoader";
 import "./JSONLoader";
@@ -18,8 +17,8 @@ import "./TextureCubeLoader";
 import "./AnimationClipLoader";
 
 export { parseSingleKTX } from "./compressed-texture";
-export { GLTFResource } from "./gltf/GLTFResource";
 export type { GLTFParams } from "./GLTFLoader";
 export * from "./resource-deserialize";
 export * from "./SceneLoader";
 export type { Texture2DParams } from "./Texture2DLoader";
+export * from "./gltf";
